pymysqlhostexporter/requirements.txt
mysql-connector-python
prometheus_client